+++ src/data/source/implementation/GroupRankDataSource.php
@@ -18,9 +18,11 @@
  * @file
  */
 
-namespace MediaWiki\Extension\RobloxAPI\data\source;
+namespace MediaWiki\Extension\RobloxAPI\data\source\implementation;
 
 use MediaWiki\Extension\RobloxAPI\data\args\ArgumentSpecification;
+use MediaWiki\Extension\RobloxAPI\data\source\DataSourceProvider;
+use MediaWiki\Extension\RobloxAPI\data\source\DependentDataSource;
 use MediaWiki\Extension\RobloxAPI\util\RobloxAPIException;
 use Parser;
